refactor(BA-2969, BA-2970): Extract DB source layer from container registry repository
Decouple SQLAlchemy query logic into ContainerRegistryDBSource, making
the repository a thin delegation wrapper with resilience decorators.
Move post-update validation (ContainerRegistryValidator) from repository
to service layer per architectural guidelines.
