+++ src/ai/backend/manager/repositories/container_registry/db_source/__init__.py
@@ -1,0 +1,3 @@
+from .db_source import ContainerRegistryDBSource
+
+__all__ = ("ContainerRegistryDBSource",)
