+++ src/ai/backend/manager/repositories/container_registry/repositories.py
@@ -1,6 +1,9 @@
 from dataclasses import dataclass
 from typing import Self
 
+from ai.backend.manager.repositories.container_registry.db_source.db_source import (
+    ContainerRegistryDBSource,
+)
 from ai.backend.manager.repositories.container_registry.repository import (
     ContainerRegistryRepository,
 )
@@ -13,4 +16,5 @@ class ContainerRegistryRepositories:
 
     @classmethod
     def create(cls, args: RepositoryArgs) -> Self:
-        return cls(repository=ContainerRegistryRepository(args.db))
+        db_source = ContainerRegistryDBSource(args.db)
+        return cls(repository=ContainerRegistryRepository(db_source))
