+++ tests/component/resource_preset/conftest.py
@@ -26,6 +26,9 @@ from ai.backend.manager.models.group import GroupRow
 from ai.backend.manager.models.resource_preset.row import ResourcePresetRow
 from ai.backend.manager.models.utils import ExtendedAsyncSAEngine
 from ai.backend.manager.repositories.agent.repository import AgentRepository
+from ai.backend.manager.repositories.container_registry.db_source.db_source import (
+    ContainerRegistryDBSource,
+)
 from ai.backend.manager.repositories.container_registry.repository import (
     ContainerRegistryRepository,
 )
@@ -65,7 +68,8 @@ def _create_mock_validators() -> MagicMock:
 def container_registry_processors(
     database_engine: ExtendedAsyncSAEngine,
 ) -> ContainerRegistryProcessors:
-    repo = ContainerRegistryRepository(database_engine)
+    db_source = ContainerRegistryDBSource(database_engine)
+    repo = ContainerRegistryRepository(db_source)
     service = ContainerRegistryService(database_engine, repo)
     return ContainerRegistryProcessors(
         service=service, action_monitors=[], validators=_create_mock_validators()
