+++ tests/unit/manager/repositories/container_registry/test_container_registry_repository.py
@@ -57,6 +57,9 @@ from ai.backend.manager.repositories.base.updater import Updater
 from ai.backend.manager.repositories.container_registry.creators import (
     ContainerRegistryCreatorSpec,
 )
+from ai.backend.manager.repositories.container_registry.db_source.db_source import (
+    ContainerRegistryDBSource,
+)
 from ai.backend.manager.repositories.container_registry.repository import (
     ContainerRegistryRepository,
 )
@@ -144,7 +147,8 @@ class TestContainerRegistryRepository:
     @pytest.fixture
     def repository(self, db_with_cleanup: ExtendedAsyncSAEngine) -> ContainerRegistryRepository:
         """Create ContainerRegistryRepository instance with real database"""
-        return ContainerRegistryRepository(db=db_with_cleanup)
+        db_source = ContainerRegistryDBSource(db_with_cleanup)
+        return ContainerRegistryRepository(db_source)
 
     @pytest.fixture
     async def sample_domain(self, db_with_cleanup: ExtendedAsyncSAEngine) -> str:
@@ -1343,7 +1347,8 @@ class TestSearchContainerRegistries:
 
     @pytest.fixture
     def repository(self, db_with_cleanup: ExtendedAsyncSAEngine) -> ContainerRegistryRepository:
-        return ContainerRegistryRepository(db=db_with_cleanup)
+        db_source = ContainerRegistryDBSource(db_with_cleanup)
+        return ContainerRegistryRepository(db_source)
 
     @pytest.fixture
     async def sample_registries(
